+++ Changelog.md
@@ -16,6 +16,7 @@ Bugfixes:
  * Inline assembly: fix parsing of assignment after a label.
  * Inline assembly: external variables of unsupported type (such as ``this``, ``super``, etc.)
    are properly detected as unusable.
+ * Inline assembly: support variables within modifiers.
 
 ### 0.4.2 (2016-09-17)